requirements.txt
tensorflow==2.13.0
transformers==4.35.0
pandas==2.1.0
pyarrow==14.0.0
numpy==1.24.3
scikit-learn==1.3.0
pyyaml==6.0
tqdm==4.66.1
fsspec[hdfs]==2023.10.0
tensorflow-addons==0.21.0
keras-tuner==1.4.0
pyspark==3.5.0
